# swagger.yml
openapi: 3.0.0
info:
  title: Minecraft Forum API
  description: An API for Minecraft players to ask questions, chat with the community or share content
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
...
    SubjectRequestBody:
      description: Request body for creating or updating a subject
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                description: The name of the subject
                example: "My subject"
              text:
                type: string
                description: The content of the subject
                example: "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
...
  responses:
...
    InternalServerError:
      description: An error occurred on the server.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: A message describing the error.
                example: "An unexpected error occurred on the server."
    BadRequestError:
      description: The request was invalid or could not be understood by the server.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: A message describing the error.
                example: "Invalid request body."
    NotFoundError:
      description: The requested resource was not found on the server.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: A message describing the error.
                example: "The requested section was not found."
  schemas:
...
    Subject:
      type: object
      properties:
        id:
          type: string
          example: "50c42313-4efc-4f57-aa9e-e2d3a3a69a2f"
        name:
          type: string
          example: "How to make a redstone contraption?"
        section_id:
          type: string
          example: "37752b5d-06e9-4c6d-afca-820db6d12f6d"
        author_id:
          type: string
          example: "8a665d4e-56cb-4ec4-9d4c-ecf5c1d79df5"
        created_at:
          type: string
          format: date-time
          example: "2023-03-04T16:53:47.685Z"
      required:
        - id
        - name
        - section_id
        - author_id
        - created_at
...
  parameters:
    SectionIdParameter:
      name: sectionId
      in: path
      required: true
      description: The ID of the section to retrieve.
      schema:
        type: string
        format: uuid
    SubjectIdParameter:
      name: subjectId
      in: path
      description: The ID of the subject to delete
      required: true
      schema:
        type: integer
        format: int64
...
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "user123"
                password:
                  type: string
                  format: password
                  example: "mysecretpassword"
              required:
                - username
                - password
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid request body
        '409':
          description: User with the same username already exists
      tags:
        - Auth
  /auth/login:
    post:
      summary: Authenticate a user and obtain a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "user123"
                password:
                  type: string
                  format: password
                  example: "mysecretpassword"
              required:
                - username
                - password
      responses:
        '200':
          description: JWT token obtained successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiODg4NzIzNzAtNjkzOC00NjExLWFiMDQtYjRhZmRiNTI4MjhlIiwiaWF0IjoxNjIzNDIxNzA0LCJleHAiOjE2MjM0MjUzMDR9.5OoFonrEh0fVl_1bGQ2Y7MjGvQ8Wk5LRzpq5R5eBN-E"
        '401':
          description: Invalid username or password
      tags:
        - Auth
...
  /sections/{sectionId}/subjects:
    get:
      tags:
        - subjects
      summary: Get all subjects for a section
      description: Retrieves all subjects that belong to a given section.
      operationId: getSubjectsForSection
      parameters:
        - $ref: '#/components/parameters/SectionIdParameter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - subjects
      summary: Create a new subject in a section
      security:
        - bearerAuth: []
      description: Creates a new subject within a given section.
      operationId: createSubjectInSection
      parameters:
        - $ref: '#/components/parameters/SectionIdParameter'
      requestBody:
        $ref: '#/components/requestBodies/SubjectRequestBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sections/{sectionId}/subjects/{subjectId}:
    get:
      tags:
        - subjects
      summary: Get a subject in a section
      description: Retrieves a single subject that belongs to a given section.
      operationId: getSubjectInSection
      parameters:
        - $ref: '#/components/parameters/SectionIdParameter'
        - $ref: '#/components/parameters/SubjectIdParameter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - subjects
      summary: Update a subject in a section
      security:
        - bearerAuth: []
      description: Updates a single subject that belongs to a given section.
      operationId: updateSubjectInSection
      parameters:
        - $ref: '#/components/parameters/SectionIdParameter'
        - $ref: '#/components/parameters/SubjectIdParameter'
      requestBody:
        $ref: '#/components/requestBodies/SubjectRequestBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - subjects
      summary: Delete a subject from a section
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SectionIdParameter'
        - $ref: '#/components/parameters/SubjectIdParameter'
      responses:
        '204':
          description: The subject was deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...